Revert "Ft search news and stocks"
