+++ src/actions/actionTypes.js
@@ -1,24 +1,12 @@
 export const IS_LOADING = 'IS_LOADING';
-
 export const LOAD_ACTIVE_USER_SUCCESS = 'LOAD_ACTIVE_USER_SUCCESS';
-
 export const FETCH_FEATURED_NEWS_SUCCESS = 'FETCH_FEATURED_NEWS_SUCCESS';
-export const FETCH_FEATURED_NEWS_FAILURE = 'FETCH_FEATURED_NEWS_FAILURE';
-
 export const FETCH_GENERAL_NEWS_SUCCESS = 'FETCH_GENERAL_NEWS_SUCCESS';
+export const FETCH_FEATURED_NEWS_FAILURE = 'FETCH_FEATURED_NEWS_FAILURE';
 export const FETCH_GENERAL_NEWS_FAILURE = 'FETCH_GENERAL_NEWS_FAILURE';
-export const FETCH_GENERAL_NEWS_LOADING = 'FETCH_GENERAL_NEWS_LOADING';
-
 export const FETCH_WORLD_NEWS_SUCCESS = 'FETCH_WORLD_NEWS_SUCCESS';
 export const FETCH_WORLD_NEWS_FAILURE = 'FETCH_WORLD_NEWS_FAILURE';
-
+export const FETCH_GENERAL_NEWS_LOADING = 'FETCH_WORLD_NEWS_LOADING';
 export const FETCH_FEATURED_NEWS_LOADING = 'FETCH_FEATURED_NEWS_LOADING';
-
 export const FETCH_INSIDE_BUSINESS_LOADING = 'FETCH_INSIDE_BUSINESS_LOADING';
 export const FETCH_INSIDE_BUSINESS_SUCCESS = 'FETCH_INSIDE_BUSINESS_SUCCESS';
-
-export const SEARCH_NEWS_SUCCESS = 'SEARCH_NEWS_SUCCESS';
-export const SEARCH_NEWS_FAILURE = 'SEARCH_NEWS_FAILURE';
-
-export const SEARCH_STOCKS_SUCCESS = 'SEARCH_STOCKS_SUCCESS';
-export const SEARCH_STOCKS_FAILURE = 'SEARCH_STOCKS_FAILURE';
