+++ src/config/index.js
@@ -6,17 +6,18 @@ const allConfig = {
 
     production: {
         BASE_URL: 'https://stockman-app.herokuapp.com',
-        STOCKMAN_API_BASE_URL: 'http://stockmanapi-env.fptfzc4bbf.us-east-2.elasticbeanstalk.com/api/v1',
+        STOCKMAN_API_BASE_URL: 'https://api.stockman-app.herokuapp.com/api/v1',
     },
 
     staging: {
         BASE_URL: 'https://staging-stockman-app.herokuapp.com',
-        STOCKMAN_API_BASE_URL: 'http://stockmanapi-env.fptfzc4bbf.us-east-2.elasticbeanstalk.com/api/v1',
+        STOCKMAN_API_BASE_URL:
+            'https://staging.stockman-app.herokuapp.com/api/v1',
     },
 
     test: {
         BASE_URL: 'http://localhost:3000',
-        STOCKMAN_API_BASE_URL: 'http://stockmanapi-env.fptfzc4bbf.us-east-2.elasticbeanstalk.com/api/v1',
+        STOCKMAN_API_BASE_URL: 'http://localhost:8000/api/v1',
     },
 };
